# Conflicts:
#	src/META-INF/plugin.xml

+++ src/com/dci/intellij/dbn/common/thread/ThreadProperty.java
@@ -23,10 +23,10 @@ public enum ThreadProperty implements Property{
         return propagatable;
     }
 
-    private final int index = Property.idx(this);
+    private final long index = Property.idx(this);
 
     @Override
-    public int index() {
+    public long index() {
         return index;
     }
 }
